change all decimal to float, add stock, helpful

+++ db/migrate/20211012190517_create_gift_cards.rb
@@ -2,7 +2,7 @@ class CreateGiftCards < ActiveRecord::Migration[6.1]
   def change
     create_table :gift_cards do |t|
       t.bigint :code
-      t.decimal :balance
+      t.float :balance
       t.string :notes
       t.belongs_to :user
