+++ db/migrate/20211012191106_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[6.1]
   def change
     create_table :orders do |t|
-      t.decimal :total
+      t.float :total
       t.string :ship_to_address
       t.string :status
       t.date :order_placed
