+++ db/migrate/20211012191511_create_shopping_carts.rb
@@ -2,7 +2,7 @@ class CreateShoppingCarts < ActiveRecord::Migration[6.1]
   def change
     create_table :shopping_carts do |t|
       t.integer :item_count, default: 0
-      t.decimal :total
+      t.float :total
       t.belongs_to :user, null: true
 
       t.timestamps
